table buttons refactor
