+++ app/react/V2/Components/UI/Table/Table.tsx
@@ -9,6 +9,7 @@ import {
   SortingState,
   getSortedRowModel,
   RowSelectionState,
+  ExpandedState,
 } from '@tanstack/react-table';
 import {
   DragEndEvent,
@@ -29,6 +30,7 @@ import { SortingChevrons } from './SortingChevrons';
 import { GroupCell, GroupHeader } from './GroupComponents';
 import { NoDataRow } from './NoDataRow';
 import { DefaultNoDataMessage } from './DefaultNoDataMessage';
+import { Button } from '../Button';
 
 type TableRow<T> = {
   rowId: string;
@@ -58,30 +60,29 @@ type TableProps<T extends TableRow<T>> = {
   defaultSorting?: SortingState;
   sortingFn?: (sorting: SortingState) => void;
   header?: React.ReactNode;
+  actions?: React.ReactNode;
   footer?: React.ReactNode;
   noDataMessage?: string | React.ReactNode;
   className?: string;
   groupColumnPosition?: number;
 };
 
-const Table = <T extends TableRow<T>>(
-  {
-    columns,
-    data,
-    onChange,
-    dnd,
-    enableSelections,
-    defaultSorting,
-    sortingFn,
-    header,
-    footer,
-    className,
-    noDataMessage = <DefaultNoDataMessage />,
-    groupColumnPosition = 0,
-    initialSelection = [],
-  }: TableProps<T>,
-  ref: React.Ref<TableRef>
-) => {
+const Table = <T extends TableRow<T>>({
+  columns,
+  data,
+  onChange,
+  dnd,
+  enableSelections,
+  defaultSorting,
+  sortingFn,
+  header,
+  actions,
+  footer,
+  className,
+  noDataMessage = <DefaultNoDataMessage />,
+  groupColumnPosition = 0,
+  initialSelection = [],
+}: TableProps<T>) => {
   const [dataState, setDataState] = useState(data);
   const initialRowSelection = initialSelection.reduce(
     (acc, item) => ({ ...acc, [item.rowId]: true }),
@@ -89,6 +90,7 @@ const Table = <T extends TableRow<T>>(
   );
   const [rowSelection, setRowSelection] = useState<RowSelectionState>(initialRowSelection);
   const [sortingState, setSortingState] = useState<SortingState>(defaultSorting || []);
+
   const rowIds = useMemo(() => getRowIds(dataState), [dataState]);
   const { memoizedColumns, groupColumnIndex } = useMemo<{
     memoizedColumns: ColumnDef<T, any>[];
@@ -174,9 +176,7 @@ const Table = <T extends TableRow<T>>(
 
   const collapseAll = () => {
     table.getRowModel().rows.forEach(row => {
-      if (row.getCanExpand()) {
-        row.toggleExpanded(false);
-      }
+      row.toggleExpanded(false);
     });
   };
 
@@ -187,32 +187,6 @@ const Table = <T extends TableRow<T>>(
       }
     });
   };
-
-  // --- Expansion state values ---
-  const [canExpand, setCanExpand] = useState(false);
-  const [canCollapse, setCanCollapse] = useState(false);
-
-  // Compute expansion state string for effect dependency
-  const expandedState = table
-    .getRowModel()
-    .rows.map(row => row.getIsExpanded())
-    .join(',');
-
-  useEffect(() => {
-    setCanExpand(table.getRowModel().rows.some(row => row.getCanExpand() && !row.getIsExpanded()));
-    setCanCollapse(table.getRowModel().rows.some(row => row.getCanExpand() && row.getIsExpanded()));
-  }, [expandedState]);
-
-  useImperativeHandle(
-    ref,
-    () => ({
-      collapseAll,
-      expandAll,
-      canExpand,
-      canCollapse,
-    }),
-    [canExpand, canCollapse]
-  );
 
   const handleDragEnd = (event: DragEndEvent) => {
     const { active, over } = event;
@@ -247,6 +221,12 @@ const Table = <T extends TableRow<T>>(
     useSensor(KeyboardSensor, {})
   );
 
+  const hasGroups = dataState.some(row => row.subRows);
+  const canExpand = table
+    .getRowModel()
+    .rows.some(row => row.getCanExpand() && !row.getIsExpanded());
+  const canCollapse = table.getRowModel().rows.some(row => row.getIsExpanded());
+
   return (
     <DndContext
       collisionDetection={closestCenter}
@@ -255,8 +235,23 @@ const Table = <T extends TableRow<T>>(
       sensors={sensors}
     >
       <div className="w-full overflow-auto rounded-md shadow">
+        <div className="flex justify-between items-center p-4">
+          {header && <caption>{header}</caption>}
+          <div className="flex gap-2">
+            {hasGroups && (
+              <>
+                <Button disabled={!canCollapse} styling="light" onClick={collapseAll}>
+                  Collapse all
+                </Button>
+                <Button disabled={!canExpand} styling="light" onClick={expandAll}>
+                  Expand all
+                </Button>
+              </>
+            )}
+            {actions}
+          </div>
+        </div>
         <table className={`w-full ${className || ''}`}>
-          {header && <caption className="p-4">{header}</caption>}
           <thead className="bg-gray-50">
             {table.getHeaderGroups().map(headerGroup => (
               <tr key={headerGroup.id}>
@@ -308,10 +303,5 @@ const Table = <T extends TableRow<T>>(
   );
 };
 
-// Wrap Table with forwardRef and export
-const ForwardedTable = React.forwardRef(Table) as <T extends TableRow<T>>(
-  props: TableProps<T> & { ref?: React.Ref<TableRef> }
-) => ReturnType<typeof Table>;
-
 export type { TableProps, TableRow, TableRef };
-export { ForwardedTable as Table };
+export { Table };
